@@ -1,351 +1,159 @@
 package com.methk.robocar;
 
-import android.app.Dialog;
-import android.content.BroadcastReceiver;
+import android.Manifest;
 import android.content.Context;
-import android.content.Intent;
-import android.content.IntentFilter;
 import android.content.SharedPreferences;
-import android.content.res.Configuration;
-import android.hardware.Sensor;
-import android.hardware.SensorManager;
+import android.content.pm.PackageManager;
+import android.location.Location;
 import android.os.Bundle;
-import android.support.design.widget.Snackbar;
-import android.support.v4.content.LocalBroadcastManager;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
-import android.widget.EditText;
-import android.widget.RadioButton;
-import android.widget.RadioGroup;
-import android.widget.TextView;
 import android.widget.Toast;
 
-import java.util.List;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
 
-public class MapActivity extends AppCompatActivity implements View.OnClickListener {
+import com.google.android.gms.location.FusedLocationProviderClient;
+import com.google.android.gms.location.LocationServices;
+import com.google.android.gms.maps.CameraUpdateFactory;
+import com.google.android.gms.maps.GoogleMap;
+import com.google.android.gms.maps.OnMapReadyCallback;
+import com.google.android.gms.maps.SupportMapFragment;
+import com.google.android.gms.maps.model.LatLng;
+import com.google.android.gms.maps.model.MarkerOptions;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.android.gms.tasks.Task;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
-    // Widgets
-    private TextView forwardSetting, rightSetting, backwardSetting, leftSetting;
-    private TextView forwardRange, rightRange, backwardRange, leftRange;
-    private TextView forwardValue, rightValue, backwardValue, leftValue;
-    private EditText minValue, maxValue, valueChosen;
+import java.util.HashSet;
+import java.util.Set;
 
-    // Dialogs
-    private Dialog sensorsDialog, rangeDialog, valueDialog;
-    private List<Sensor> sensorsList;
-    private RadioGroup radioGroup;
+public class MapActivity extends AppCompatActivity implements OnMapReadyCallback {
 
-    // Settings
-    private int[] settings; // Array with the position of each sensor in the sensors list
-    private String[] ranges; // Array with the ranges of validity for each sensor
-    private int[] values; // Array with the valid axis for each sensor
-    private int selectedSetting = 0;
+    private static final String PREFS_NAME = "LocationPrefs";
+    private static final String KEY_LOCATIONS = "locations";
 
-    // Preferences
-    protected static final String PREFERENCES_NAME = "sensors";
-    private SharedPreferences preferences;
+    private GoogleMap mMap;
+    private FusedLocationProviderClient fusedLocationProviderClient;
+    private static final int LOCATION_PERMISSION_REQUEST_CODE = 1;
+    private Location currentLocation;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_map);
-        setSupportActionBar((Toolbar)findViewById(R.id.toolbar));
+
+        setSupportActionBar(findViewById(R.id.toolbar));
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
 
-        LocalBroadcastManager.getInstance(this).registerReceiver(new BroadcastReceiver() {
-            @Override public void onReceive(Context context, Intent intent) { finish(); } // Stop activity if bluetooth is disconnected
-        }, new IntentFilter(BluetoothConnection.INTENT_FILTER_STOP));
+        fusedLocationProviderClient = LocationServices.getFusedLocationProviderClient(this);
 
-        // Setup dialogs
-        sensorsDialog = new Dialog(this);
-        sensorsDialog.setCancelable(false);
-        rangeDialog = new Dialog(this);
-        rangeDialog.setCancelable(false);
-        valueDialog = new Dialog(this);
-        valueDialog.setCancelable(false);
-
-        // Initialize widgets
-        initWidgets();
-
-        preferences = getSharedPreferences(PREFERENCES_NAME, MODE_PRIVATE);
-        try {
-            sensorsList = ((SensorManager) getSystemService(SENSOR_SERVICE)).getSensorList(Sensor.TYPE_ALL);
-        } catch(NullPointerException e) {
-            Toast.makeText(getApplicationContext(), getString(R.string.settings_problem_list), Toast.LENGTH_LONG).show();
-            finish();
+        SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager()
+                .findFragmentById(R.id.map);
+        if (mapFragment != null) {
+            mapFragment.getMapAsync(this);
         }
 
-        // Setup sensors dialog content
-        for(int i = 0; i < sensorsList.size(); i++) {
-            RadioButton radioButton = new RadioButton(this);
-            radioButton.setText(sensorsList.get(i).getName());
-            radioButton.setId(i);
-            radioGroup.addView(radioButton);
-        }
-
-        // Initialize settings
-        settings = new int[4]; ranges = new String[4]; values = new int[4];
-        for(int i = 0; i < 4; i++) {
-            if(preferences.getInt("setting" + i, -1) == -1) { // If a value was not saved then set default settings
-                if(setStandardSettings()) break; // If standard sensors are available
-            } else
-                settings[i] = preferences.getInt("setting" + i, 0);
-            ranges[i] = preferences.getString("range" + i, "0:0");
-            values[i] = preferences.getInt("value" + i, 0);
-
-        }
-
-        updateSettingsLabels();
-        updateRangesLabels();
-        updateValuesLabels();
-    }
-
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-    }
-
-    @Override
-    public boolean onSupportNavigateUp() {
-        finish();
-        return true;
-    }
-
-    @Override
-    public void onBackPressed() {
-        finish();
-    }
-
-    @Override
-    public void onConfigurationChanged(Configuration newConfig) {
-        super.onConfigurationChanged(newConfig);
-        setContentView(R.layout.activity_settings);
-        initWidgets();
-    }
-
-    public void initWidgets() {
-        findViewById(R.id.fab).setOnClickListener(this);
-
-        forwardSetting = findViewById(R.id.tw_fs);
-        forwardSetting.setOnClickListener(this);
-        rightSetting = findViewById(R.id.tw_rs);
-        rightSetting.setOnClickListener(this);
-        backwardSetting = findViewById(R.id.tw_bs);
-        backwardSetting.setOnClickListener(this);
-        leftSetting = findViewById(R.id.tw_ls);
-        leftSetting.setOnClickListener(this);
-
-        forwardRange = findViewById(R.id.tw_fr);
-        forwardRange.setOnClickListener(this);
-        rightRange = findViewById(R.id.tw_rr);
-        rightRange.setOnClickListener(this);
-        backwardRange = findViewById(R.id.tw_br);
-        backwardRange.setOnClickListener(this);
-        leftRange = findViewById(R.id.tw_lr);
-        leftRange.setOnClickListener(this);
-
-        forwardValue = findViewById(R.id.tw_fp);
-        forwardValue.setOnClickListener(this);
-        rightValue = findViewById(R.id.tw_rp);
-        rightValue.setOnClickListener(this);
-        backwardValue = findViewById(R.id.tw_bp);
-        backwardValue.setOnClickListener(this);
-        leftValue = findViewById(R.id.tw_lp);
-        leftValue.setOnClickListener(this);
-
-        findViewById(R.id.btn_standard).setOnClickListener(this);
-
-        // Setup dialogs layouts
-        sensorsDialog.setContentView(R.layout.radiobuttons_dialog);
-        rangeDialog.setContentView(R.layout.range_dialog);
-        valueDialog.setContentView(R.layout.value_dialog);
-
-        radioGroup = sensorsDialog.findViewById(R.id.radio_group);
-        sensorsDialog.findViewById(R.id.btn_close).setOnClickListener(this);
-
-        rangeDialog.findViewById(R.id.btn_close2).setOnClickListener(this);
-        minValue = rangeDialog.findViewById(R.id.et_min);
-        maxValue = rangeDialog.findViewById(R.id.et_max);
-
-        valueDialog.findViewById(R.id.btn_close3).setOnClickListener(this);
-        valueChosen = valueDialog.findViewById(R.id.et_val);
-    }
-
-    private boolean setStandardSettings() {
-        int proximitySensorIndex = -1, gameRotationSensorIndex = -1;
-
-        for(int i = 0; i < sensorsList.size(); i++) { // Search for proximity and rotation sensors
-            if(proximitySensorIndex == -1 && sensorsList.get(i).getType() == Sensor.TYPE_PROXIMITY && sensorsList.get(i).getMaximumRange() > 0.0) {
-                proximitySensorIndex = i;
-            } else if (gameRotationSensorIndex == -1 && sensorsList.get(i).getType() == Sensor.TYPE_GAME_ROTATION_VECTOR && sensorsList.get(i).getMaximumRange() > 0.0) {
-                gameRotationSensorIndex = i;
+        FloatingActionButton backButton = findViewById(R.id.back_button);
+        backButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                onBackPressed();
             }
-        }
+        });
 
-        if(proximitySensorIndex > -1 && gameRotationSensorIndex > -1) {
-            settings[0] = proximitySensorIndex;
-            ranges[0] = "0.0:0.5";
-            values[0] = 0;
+        FloatingActionButton saveButton = findViewById(R.id.save_button);
+        saveButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                saveLocation();
+            }
+        });
 
-            settings[1] = gameRotationSensorIndex;
-            ranges[1] = "0.0:0.25";
-            values[1] = 0;
+        FloatingActionButton locationButton = findViewById(R.id.location_button);
+        locationButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                getCurrentLocation();
+            }
+        });
+    }
 
-            settings[2] = proximitySensorIndex;
-            ranges[2] = "0.5:1.0";
-            values[2] = 0;
-
-            settings[3] = gameRotationSensorIndex;
-            ranges[3] = "0.75:1.0";
-            values[3] = 0;
-
-            updateSettingsLabels();
-            updateRangesLabels();
-            updateValuesLabels();
-
-            savePreferences();
-            return true;
+    @Override
+    public void onMapReady(GoogleMap googleMap) {
+        mMap = googleMap;
+        if (ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED) {
+            mMap.setMyLocationEnabled(true);
         } else {
-            Toast.makeText(this, getString(R.string.settings_problem_standard), Toast.LENGTH_LONG).show();
-            return false;
+            ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.ACCESS_FINE_LOCATION}, LOCATION_PERMISSION_REQUEST_CODE);
         }
+        loadSavedLocations();
     }
 
-    private void updateSettingsLabels() {
-        forwardSetting.setText(sensorsList.get(settings[0]).getName().length() > 16 ? sensorsList.get(settings[0]).getName().substring(0, 16) + "..." : sensorsList.get(settings[0]).getName());
-        rightSetting.setText(sensorsList.get(settings[1]).getName().length() > 16 ? sensorsList.get(settings[1]).getName().substring(0, 16) + "..." : sensorsList.get(settings[1]).getName());
-        backwardSetting.setText(sensorsList.get(settings[2]).getName().length() > 16 ? sensorsList.get(settings[2]).getName().substring(0, 16) + "..." : sensorsList.get(settings[2]).getName());
-        leftSetting.setText(sensorsList.get(settings[3]).getName().length() > 16 ? sensorsList.get(settings[3]).getName().substring(0, 16) + "..." : sensorsList.get(settings[3]).getName());
-    }
-
-    private void updateRangesLabels() {
-        forwardRange.setText(ranges[0].length() > 16 ? ranges[0].substring(0, 16) + "..." : ranges[0]);
-        rightRange.setText(ranges[1].length() > 16 ? ranges[1].substring(0, 16) + "..." : ranges[1]);
-        backwardRange.setText(ranges[2].length() > 16 ? ranges[2].substring(0, 16) + "..." : ranges[2]);
-        leftRange.setText(ranges[3].length() > 16 ? ranges[3].substring(0, 16) + "..." : ranges[3]);
-    }
-
-    private void updateValuesLabels() {
-        forwardValue.setText(String.valueOf(values[0]));
-        rightValue.setText(String.valueOf(values[1]));
-        backwardValue.setText(String.valueOf(values[2]));
-        leftValue.setText(String.valueOf(values[3]));
-    }
-
-    private void showSensorsDialog() {
-        radioGroup.check(radioGroup.getChildAt(settings[selectedSetting]).getId());
-        sensorsDialog.show();
-    }
-
-    private void showRangeDialog() {
-        String[] values = ranges[selectedSetting].split(":");
-        minValue.setText(values[0]);
-        maxValue.setText(values[1]);
-
-        rangeDialog.show();
-    }
-
-    private void showValueDialog() {
-        valueChosen.setText(String.valueOf(values[selectedSetting]));
-        valueDialog.show();
+    private void getCurrentLocation() {
+        if (ActivityCompat.checkSelfPermission(MapActivity.this, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(MapActivity.this, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
+            ActivityCompat.requestPermissions(MapActivity.this, new String[]{Manifest.permission.ACCESS_FINE_LOCATION}, LOCATION_PERMISSION_REQUEST_CODE);
+            return;
+        }
+        Task<Location> task = fusedLocationProviderClient.getLastLocation();
+        task.addOnSuccessListener(MapActivity.this, new OnSuccessListener<Location>() {
+            @Override
+            public void onSuccess(Location location) {
+                if (location != null) {
+                    currentLocation = location;
+                    LatLng latLng = new LatLng(location.getLatitude(), location.getLongitude());
+                    mMap.addMarker(new MarkerOptions().position(latLng).title("Менің орналасқан жерім"));
+                    mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(latLng, 15));
+                    if (ActivityCompat.checkSelfPermission(MapActivity.this, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(MapActivity.this, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
+                        return;
+                    }
+                    mMap.setMyLocationEnabled(true);
+                }
+            }
+        });
     }
 
     @Override
-    public void onClick(View v) {
-        switch (v.getId()) {
-            case R.id.fab:
-                savePreferences();
-                Snackbar.make(v, getString(R.string.settings_saved), Snackbar.LENGTH_LONG).show();
-                break;
-            case R.id.btn_close: // Sensors dialog
-                settings[selectedSetting] = radioGroup.getCheckedRadioButtonId();
-                updateSettingsLabels();
-                sensorsDialog.dismiss();
-                break;
-            case R.id.btn_close2: // Ranges dialog
-                if(Double.parseDouble(minValue.getText().toString()) > 1 || Double.parseDouble(maxValue.getText().toString()) > 1) {
-                    Toast.makeText(this, getString(R.string.settings_err1), Toast.LENGTH_SHORT).show();
-                } else if(Double.parseDouble(minValue.getText().toString()) > Double.parseDouble(maxValue.getText().toString())) {
-                    Toast.makeText(this, getString(R.string.settings_err2), Toast.LENGTH_SHORT).show();
-                } else {
-                    ranges[selectedSetting] = minValue.getText().toString() + ":" + maxValue.getText().toString();
-                }
-                updateRangesLabels();
-                rangeDialog.dismiss();
-                break;
-            case R.id.btn_close3: // Values dialog
-                if(Integer.parseInt(valueChosen.getText().toString()) > 5 || Integer.parseInt(valueChosen.getText().toString()) < 0)
-                    Toast.makeText(this, getString(R.string.settings_err3), Toast.LENGTH_SHORT).show();
-                else
-                    values[selectedSetting] = Integer.parseInt(valueChosen.getText().toString());
-                updateValuesLabels();
-                valueDialog.dismiss();
-                break;
-            case R.id.btn_standard:
-                setStandardSettings();
-                break;
-            case R.id.tw_fs:
-                selectedSetting = 0;
-                showSensorsDialog();
-                break;
-            case R.id.tw_rs:
-                selectedSetting = 1;
-                showSensorsDialog();
-                break;
-            case R.id.tw_bs:
-                selectedSetting = 2;
-                showSensorsDialog();
-                break;
-            case R.id.tw_ls:
-                selectedSetting = 3;
-                showSensorsDialog();
-                break;
-            case R.id.tw_fr:
-                selectedSetting = 0;
-                showRangeDialog();
-                break;
-            case R.id.tw_rr:
-                selectedSetting = 1;
-                showRangeDialog();
-                break;
-            case R.id.tw_br:
-                selectedSetting = 2;
-                showRangeDialog();
-                break;
-            case R.id.tw_lr:
-                selectedSetting = 3;
-                showRangeDialog();
-                break;
-            case R.id.tw_fp:
-                selectedSetting = 0;
-                showValueDialog();
-                break;
-            case R.id.tw_rp:
-                selectedSetting = 1;
-                showValueDialog();
-                break;
-            case R.id.tw_bp:
-                selectedSetting = 2;
-                showValueDialog();
-                break;
-            case R.id.tw_lp:
-                selectedSetting = 3;
-                showValueDialog();
-                break;
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        if (requestCode == LOCATION_PERMISSION_REQUEST_CODE) {
+            if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                getCurrentLocation();
+            } else {
+                Toast.makeText(this, "Орналасу үшін қажетті рұқсаттар.", Toast.LENGTH_SHORT).show();
+            }
         }
     }
 
-    private void savePreferences() {
-        for(int i = 0; i < 4; i++) {
-            if(preferences.getInt("setting" + i, -1) != settings[i]) { // Save sensors selected (if changed)
-                preferences.edit().putInt("setting" + i, settings[i]).apply();
-                preferences.edit().putBoolean("changedSensor" + i, true).apply(); // The sensor has been changed
-            }
-            if(!preferences.getString("range" + i, "").equals(ranges[i])) // Save range (if changed)
-                preferences.edit().putString("range" + i, ranges[i]).apply();
-            if(preferences.getInt("value" + i, -1) != values[i]) // Save value (if changed)
-                preferences.edit().putInt("value" + i, values[i]).apply();
+    private void saveLocation() {
+        if (currentLocation != null) {
+            double latitude = currentLocation.getLatitude();
+            double longitude = currentLocation.getLongitude();
+            saveToPreferences(latitude, longitude);
+            Toast.makeText(this, "Орналасқан жері сақталды: " + latitude + ", " + longitude, Toast.LENGTH_SHORT).show();
+        } else {
+            Toast.makeText(this, "Орналасқан жері анықталмаған.", Toast.LENGTH_SHORT).show();
+        }
+    }
+
+    private void saveToPreferences(double latitude, double longitude) {
+        SharedPreferences sharedPreferences = getSharedPreferences(PREFS_NAME, Context.MODE_PRIVATE);
+        Set<String> locations = sharedPreferences.getStringSet(KEY_LOCATIONS, new HashSet<String>());
+        locations.add(latitude + "," + longitude);
+        sharedPreferences.edit().putStringSet(KEY_LOCATIONS, locations).apply();
+    }
+
+    private void loadSavedLocations() {
+        SharedPreferences sharedPreferences = getSharedPreferences(PREFS_NAME, Context.MODE_PRIVATE);
+        Set<String> locations = sharedPreferences.getStringSet(KEY_LOCATIONS, new HashSet<String>());
+        for (String location : locations) {
+            String[] latLng = location.split(",");
+            double lat = Double.parseDouble(latLng[0]);
+            double lon = Double.parseDouble(latLng[1]);
+            LatLng savedLatLng = new LatLng(lat, lon);
+            mMap.addMarker(new MarkerOptions().position(savedLatLng).title("Сақталған орын"));
+            mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(savedLatLng, 10));
         }
     }
 }
